Clean up Dockerfiles
This change aims to reduce the amount of layers when building the
various Docker images.

With the availability of `atomic` in the official distribution repos, we
are able install it via `yum`/`dnf` instead of building from source.

Additionally, we can install `openscap-daemon` from EPEL and
`openscap-containers` from the OpenSCAP copr repo maintained by
@isimluk, so we can avoid building them from source as well.

The formatting of the Dockerfiles is changed to (hopefully) improve the
readability of them.