Fix 'Description' field in JSON results
The XCCDF description element usually contains child elements,
that are used for formating. To get the whole description, we cannot
use the 'text' attribute, because it contains only text before the
first child and the rest is omitted. Instead, we must use tostring()
method to serialize all the description text.