Generate remediation script after scan
We need a remediation script to be able to build hardened
images. Because we will need to use the script in Atomic,
which is independent on scanner, we can't call
"oscap xccdf generate fix" directly from Atomic.
Instead, we have to generate the fix right after the scan
by oscapd-evaluate.