Fix import cElementTree module
The xml.etree.cElementTree is deprecated since Python 3.3 and
it will be removed in Python 3.9. It can be replaced by
xml.etree.ElementTree. See https://bugs.python.org/issue36543.
We still want to import cElementTree when using Python 2,
so we will try to import it and if the import fails we will
use the latter.

Resolves: RHBZ#1817660